Refactor models and test helpers out of `main.rs`
This commit leaves a lot of `todo!()`s but sets up a more manageable
file structure for future work.

It also starts bringing in AWS dependencies to set up to make the
lambda actually do something, which was the trigger for deciding to
refactor.

I've added a unit test but both tests are also now failing while I work
out how best to test now we're calling AWS APIs and implement the
models fully.