Update `UserRecord` with the service from the event
There are two possibilities when updating a record:

1. The user hasn't used that service before, so we can just add the
new service to their list.

2. The user has previously used that service. In this case, we have to
keep the service in their list but update the last accessed value

We can achieve both of these by dropping any services from the user's
list that have the same service ID as the one in the event, then
adding a new one.