Remove unneeded test
This test could never pass as `process_message` requires the AWS config
to be in the environment and calls out to the real DynamoDB service.

Rust doesn't mock these calls for testing in the same way as something
like Ruby or Python, so simply remove this test while I work out the
correct way of testing things.

We're not getting a lot of value out of this test even if it did work -
most of the logic gets pushed down to the models by the traits which
all have their own tests anyway.