Add SAM template and deploy script
This template sets up what's needed to demo the event handler Lambda.
It creates an SQS queue, a DynamoDB table, the Lambda and the
permissions for the Lambda to listen to the queue and read/write to
DynamoDB.

I've also added a deploy script to remind me what to do.